% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createclass.R
\name{createclass}
\alias{createclass}
\title{Read delimited text file into sclass-object}
\usage{
createclass(pathStr, fileStr)
}
\arguments{
\item{pathStr}{input path to MaxQuant results as string}

\item{fileStr}{input filename of MaxQuant results as string}
}
\description{
Read delimited text file into sclass-object
}
