% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterlfq.R
\name{filterlfq}
\alias{filterlfq}
\title{Removes protein groups with a certain number of missing values}
\usage{
filterlfq(mqo, maxTotNA)
}
\arguments{
\item{mqo}{input maxquant object}

\item{maxTotNA}{input, maximum number of missing values in each lfq row}
}
\description{
Removes protein groups with a certain number of missing values
}
