% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getlfq.R
\name{getlfq}
\alias{getlfq}
\title{Extract LFQ intensity matrix from proteinGroups and put in slot LFQ}
\usage{
getlfq(mqo)
}
\arguments{
\item{mqo}{input maxquant object}
}
\description{
Extract LFQ intensity matrix from proteinGroups and put in slot LFQ
}
